// frontend/src/api-generated/models/cauHinhDonViResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface CauHinhDonViResponse {
  id?: number;
  idDonVi?: number;
  idDanhMucCauHinh?: number;
  maCauHinh?: string;
  tenCauHinh?: string;
  giaTriCauHinh?: string;
}